moved SimuComWorkflowConfiguration to de.uka.ipd.sdq.codegen.simucontroller.core
