+++ bundles/de.uka.ipd.sdq.simucom.rerunsimulation/src/de/uka/ipd/sdq/simucom/rerunsimulation/runconfig/RerunSimuComWorkflowConfiguration.java
@@ -2,12 +2,12 @@ package de.uka.ipd.sdq.simucom.rerunsimulation.runconfig;
 
 import java.util.Map;
 
-import de.uka.ipd.sdq.codegen.simucontroller.runconfig.SimuComWorkflowConfiguration;
+import de.uka.ipd.sdq.codegen.simucontroller.core.runconfig.SimuComWorkflowConfiguration;
 
 /**
  * Configuration class for the rerun simulation plug-in. Extends from
  * 
- * @see de.uka.ipd.sdq.codegen.simucontroller.runconfig.SimuComWorkflowConfiguration
+ * @see de.uka.ipd.sdq.codegen.simucontroller.core.runconfig.SimuComWorkflowConfiguration
  * 
  *      and provides information on which parts of the generated code should be regenerated.
  *      Possible parts are allocation, system and usage.
